requirements.txt
colorama==0.4.6
greenlet==3.3.0
iniconfig==2.3.0
joblib==1.5.2
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.3.251201
pluggy==1.6.0
Pygments==2.19.2
pyodbc==5.3.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
SQLAlchemy==2.0.45
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2